requirements.txt
asgiref==3.11.0
attrs==25.4.0
Babel==2.14.0
coverage==7.13.4
currency.converter==0.5.5
CurrencyConverter==0.18.14
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.4.0
freezegun==1.5.5
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
loguru==0.7.3
packaging==26.0
pluggy==1.6.0
pycountry==22.3.5
Pygments==2.19.2
pyinstrument==5.1.2
PyJWT==2.11.0
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
